the mandatory part is now finished !

+++ mandatory/include/render.h
@@ -12,7 +12,7 @@ typedef struct s_player	t_player;
 void	render_floor(t_data *data, mlx_color color);
 void	render_celling(t_data *data, mlx_color color);
 void	render_bg(t_data *data);
-void	render_walls(t_data *data, t_ray *ray, int angle);
+void	render_walls(t_data *data, t_ray *ray, float angle);
 
 void	calcul_touch(t_ray *ray, t_player *player, int dir);
 // event
